Add files via upload
The paper describes the success and failure rate of universities using data mining technique to predict the number success and Failure rate at universities. Moreover the paper describes how various attribute affects the success and failure rate. The aim of the project is to develop a model which will fulfil the early prediction using data mining algorithum which peturb the future decision. As well as the project focuses on collection large number of attributes    prognosis the future decision and will stands  as a foundation of decision support for the future forecast. The above proposal uses the decision tree and naïve baye’s algorithum.